pyproject.toml
[tool.poetry]
name = "slack-pub-bot"
version = "0.1.0"
description = ""
authors = ["bulldra <bulldra@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.3.2"
functions-framework = "^3.4.0"
google-cloud-logging = "^3.5.0"
google-cloud-pubsub = "^2.17.1"
slack_bolt = "^1.18.0"
beautifulsoup4 = "^4.12.2"
types-beautifulsoup4 = "^4.12.0"
requests = "^2.31.0"
types-requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
